Move verify routes to rmndin.users.views
