+++ rmndin/rmndin/__init__.py
@@ -21,6 +21,8 @@ jwt = JWT(app, jwt.authenticate, jwt.identity)
 
 from rmndin.reminders.views import reminders as remindersbp
 from rmndin.users.views import users as usersbp
+from rmndin.users.views import verify as verifybp
 
 app.register_blueprint(remindersbp, url_prefix='/reminders')
 app.register_blueprint(usersbp, url_prefix='/users')
+app.register_blueprint(verifybp, url_prefix='/verify')
